requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.1
transformers==4.40.1
torch>=1.13.0
email-validator==2.1.0.post1
